api errors tests: Update now that ApiError is never 5xx.
For one of these cases, there's a nice new thing the data can
be instead.

For the other, we could modify the test case to have some status
code that's neither 5xx nor 4xx.  But later in this series we're
going to eliminate all the non-4xx cases of ApiError anyway, so
we don't bother.