ios: Show a banner for notifications received while app is open.
A sound, by itself, would tell the user that something interesting
happened, but without giving them a clue about what the interesting
thing was. So, use a banner for that.

We should still eventually not show these for messages that already
appear on the screen, as we mentioned for notification sounds in the
previous commit.