sharing types [nfc]: Condense SharedData union to one definition.
This lets us not give names to the individual branches of this
union, which frees up namespace so that in particular we can use
a name like `SharedFile` for each file's data when we add support
for multiple files in `SharedData`.

This also makes it easier to take in the whole definition at once,
since it's short enough that it fits so compactly this way.