mute [nfc]: Fold shouldBeMuted the rest of the way into caller.
This puts the "what kind of narrow is this" question at the top
of the control flow for the whole getShownMessagesForNarrow
logic... which then raises some natural questions about the
various other narrows that aren't explicitly mentioned.

Add comments about those, including a TODO comment for starred
messages and the starred-messages narrow.