StreamSettingsScreen [nfc]: Convert to function component.
Several of the event handlers, like handlePressEditSubscribers, have
`useCallback`s that are currently useless, because they're invoked
in a fresh callback via `delay`.

It was likewise useless to have these event handlers defined as
properties on the instance, before this commit.

I've left the `useCallback`s in, in this commit; as Greg says [1],
this is "definitely the way to get the most direct, most clearly
no-functional-change translation of the existing pattern we usually
use on class components".

[1] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/React.2EuseCallback/near/1226585