AnimatedComponent [nfc]: Dedupe target-value expression.
It's intentional that both of these expressions do the same thing:
the point is that when the component first renders, it should start
out in the desired state, and then when the desired state changes
and causes a rerender it'll animate to the new desired state.

So, we can make that explicit (and avoid some future change causing
an accidental discrepancy) by writing the expression just once.

Also because `visible` is already boolean, we can skip the `=== true`.