msglist [nfc]: Cut redundant getMessageTransitionProps tests.
All the logic in this helper naturally belongs in its caller,
getMessageUpdateStrategy -- the logic could be significantly
simpler in that context.  The one purpose potentially served by
having it instead be a separate function is for tests.

But because the caller is the one with a meaningful interface,
the caller is also the better surface to write tests against.
And in fact the caller already has tests that cover nearly all
the same cases as the helper's tests.  So delete the latter,
converting to a getMessageUpdateStrategy test the only one of
them that isn't already covered by the existing tests there.

Then stop exporting the helper.