deps: Remove @jest/source-map.
I think we added this in fb23341c3 because Jest 25 was being
installed and used instead of Jest 26. `yarn why jest` at this
commit shows that there isn't a copy of Jest 25 lying around, so
that problem should be solved.

In any case, empirically, the "modern" fake-timer implementation is
being used successfully where we ask for it, suggesting that we're
no longer affected by the problem described in fb23341c3. The tests
run and pass, even while we're calling `jest.getRealSystemTime()` in
promiseTimeout-test.js to confirm that the "modern" implementation
is used.