i18n [nfc]: Add comment explaining `pt-BR` vs plain `pt`.
We made an explicit choice that this option wouldn't be included even
though it's above the numerical threshold, because it's redundant:
  https://github.com/zulip/zulip-mobile/pull/4901#discussion_r673472275

But without a comment in the code, as long as it's above the
numerical threshold it'll be natural for a maintainer seeing that
to go and add the option here without being aware of that reasoning.
Add such a comment.