i18n [nfc]: Explain more when to sync and to push.
We hadn't really previously written down in docs that we should be
running tx-sync after merging PRs that change the set of translated
strings.  Oops; do that.

Within the tx-sync script itself, moving the "please push promptly"
text from the usage message to a message the script prints when
applicable helps make the information visible when it's actionable
for the maintainer using the script.

Also revise and expand on some other pieces.