user: Check for logged-in account in getHaveServerData, too.
There probably isn't any UI that would actually crash as a consequence
of missing an API key for the active account.  But certainly it's
something we assume throughout the main UI, and it's easy to check
for.  So in addition to checking there's an active account at all,
check it's logged in for good measure.