register: Dispatch REGISTER_START just before we start `/register`.
With the goal that REGISTER_START and REGISTER_COMPLETE are tight
bookends around the code responsible for making the /register
request.

We hope this is NFC; the only cases where it wouldn't be are when
`getAuth` and/or `getHaveServerData` throw an error.

`getAuth` would be the natural one to worry about, but we really
don't expect `doInitialFetch` to be called unless we have an active,
logged-in account.