ChatScreen: Use fetching instead of caughtUp to refuse sending message.
If a user is scrolled up in the message list, it doesn't really matter
if caughtUp.newer is false, because the user knows they're not looking
at the latest messages anyway.  More relevant is if fetching.newer is
true, which means they're at least near the bottom.

Also the fetching condition is naturally self-limiting if we have a
working connection to the server.  The not-caughtUp condition can
persist indefinitely if we're scrolled far back in history, which
makes it a recurring frustration for users.

See chat discussion:
  https://chat.zulip.org/#narrow/stream/48-mobile/topic/Failed.20to.20send.20on.20Android/near/1158162

[greg: revised comment and commit message]