@react-navigation/stack libdef: Fix setParams by dropping $Shape.
See comment.  The underlying Flow bug is that $Shape doesn't
really work:
  https://github.com/facebook/flow/issues/5702

Happily there is a workaround, namely `$Rest<T, {...}>`.
One might give that a handy alias, like `$Partial<T>` -- and
that's exactly what this libdef already does.  So just use it here.

Also add a type-test.  The most natural place for this might be a
folder like `flow-typed/__tests__/`, next to the libdefs themselves…
but if we do that, then every time we touch the file, Flow restarts
from scratch, just as it does for actual libdefs in `flow-typed/`.
Seems like something odd there.  But we can work around it by just
using a separate folder.