example data: Fix eg.action.realm_init to agree with eg.selfUser.
Just like eg.action.login_success already does.

This covers all the properties on User that also appear in this
action, except `avatar_url` as commented.  The obstacle there is
just that `selfUser.avatar_url` is an `AvatarURL`, whereas what
this action carries is a plain `string | null`.  This is a
non-problem because we don't actually use the `avatar_url` property
on these REALM_INIT actions... and if we did, we'd probably
convert at the edge to an `AvatarURL`, just as we do for `User`,
so that that's what we'd see in the actions.