android: Enable Sentry.
This should immediately mean that we start reporting to Sentry when
an exception propagates uncaught and crashes the app.  It also means
we can start explicitly logging things to Sentry when something goes
wrong that we do anticipate and catch.

It's actually already possible to invoke Sentry, but it only works
when our JS code has already been started up, for example in a method
exposed to JS with `@ReactMethod`.  We want to use it from places like
the notifications code, which can run without starting JS.

Tested manually by following this comment's instructions with a DSN
pointing to a test "project" in Sentry (as well as the existing
instructions in `src/sentryConfig.js`), and inserting (into
`MainActivity#onCreate`) both a Sentry warning and then just an
uncaught exception.  Both were successfully reported to Sentry,
appearing in the test project's dashboard.

There is an asterisk: when I tried inducing an uncaught exception
inside a `@ReactMethod` method, I got (in a debug build) a red-screen
modal as expected, but no report to Sentry.  That seems like a Sentry
bug.  But it's a low-priority issue for us, in that AFAIK we've never
had an issue that would cause one of those.  Moreover, there looks to
be a report of it in the Sentry tracker, which says it's specific to
debug -- that it works fine in a release build.  Discussion here:
  https://github.com/zulip/zulip-mobile/pull/4996#issuecomment-916501317