outbox [nfc]: Drop spurious "whole stream as destination" case
It wouldn't make any sense to pass a whole-stream narrow here, and
indeed we don't.  The destination narrow passed to `addToOutbox`
always comes from a return value of `ComposeBox#getDestinationNarrow`,
and an assertion there already ensures that's a conversation.

The parameter name `destinationNarrow` indicates the value should come
from that method.  Tracing from the two callsites of `addToOutbox`
confirms that's the case:

 * `uploadFile` in `fetchActions`, called by `ComposeMenu`, where
   the narrow comes from its prop `destinationNarrow`.  The only
   parent of `ComposeMenu` is `ComposeBox`, which gets that prop's
   value from `getDestinationNarrow`.

 * `ChatScreen`, in its `sendCallback` callback.  That callback is
   passed only to `ComposeBox`, as its `onSend` prop; that in turn
   is invoked only at the latter component's `handleSend` method,
   where the narrow comes from `getDestinationNarrow`.

With this assertion added here, we can cheerfully drop the code
for this case (which, as the comment suspected, was already dead).
If in the future we do pass a whole-stream narrow here, we'll get
an exception, just as we already would have if passing another
inappropriate narrow like starred-messages.