compose [nfc]: Say `destinationNarrow` for narrows we're sending to
In the future, we might have a distinct type (a subtype of Narrow) for
narrows that can be used as the destination for sending a message.

At present, we don't.  But we can at least make the distinction
explicit in the names of parameters and locals.  That helps trace
the data flow, and it makes the code more readable in the context
of `ComposeBox`, where there's another narrow floating around (namely
`props.narrow`) which may be different from this one.

All of these variables named `destinationNarrow` get their values from
the return value of a call to `getDestinationNarrow`.

While we're touching the `onSend` callback prop, it also benefits
from a name to clarify the meaning of its string parameter.