account [nfc]: Mark getRealmUrl as a per-account selector
And so mark its two callers, too, because they are.

One of those, LegalScreen, is a React component, and we don't yet
have an in-code way to mark those as per-account vs. global.  But
its name kind of sounds like it might be one of the few components
of ours that *isn't* per-account -- it sounds like it could be a
global screen about legal terms related to the app itself.  So,
just add a bit of jsdoc to clarify that it's per-account after all.