event: Drop a logging call from eventToAction
This function is naturally per-account: it's all about operating
on server data, for a particular account.

We've had it calling this deprecated selector tryGetActiveAccount,
even though that naturally operates on a global state, in order to
get an early check if getCurrentRealm was going to fail.  But
because this is semantically per-account code, it's properly the
caller's job to ensure it has some per-account data to operate on.

As it happens, the caller, namely `startEventPolling`, has a bug
where it doesn't actually always manage that, which we just noted
in the previous commit.

But this check wasn't actually changing behavior -- just logging
the problem.  Now we know there is a problem, and we understand it
better than this particular logging can inform us: for example,
in addition to ceasing to have an active account at all, the user
may have simply logged out of one account.  So just drop this line.

That gives us a step toward eliminating this deprecated selector.
It'll also let us mark this function as per-account, once we deal
with getCurrentRealm which it also calls.