redux types: Tighten ThunkAction to give just PerAccountState!
This confirms that we've now explicitly marked global all the places
that we need a GlobalDispatch.  Anywhere that's still getting a
plain Dispatch -- which means the bulk of our thunk actions and
Redux components -- has to actually behave as per-account, modulo
any explicit fixmes.  (And we've only added one of those, a call to
assumeSecretlyGlobalState in eventActions.js.)