actions: Add extras getGlobalSession/Settings, and use them
This handy feature of redux-thunk lets us provide a way for
per-account thunk actions to get access to data that's legitimate
for them to have but lives outside of a PerAccountState.

We leave a fixme in the redux-mock-store libdef, with a TODO
comment so we'll know to come back to it later in this series
of PRs.  The libdef already had an imprecise type here, and
it's for a module only used in tests so it's not too critical
if there's a type error somewhere that it ends up concealing.