react-redux: Add and use connectGlobal and useGlobalDispatch
This marks as global all the places we get a `dispatch` which is
meant to be global.

For now the distinction between `connectGlobal` and plain `connect`
is only about the `dispatch` prop, not the state provided to the
`mapStateToProps` callback; so plain `connect` still provides a
GlobalState.  We'll tighten that part separately later.