settings: Don't anticipate on NotificationsScreen switches
This was basically an attempted workaround for #3999.  But it
falsely tells the user that the settings have actually taken --
there's no indication that we're still saving them.  And once
the user wises to the fact that we pretend to have saved them
when we haven't, there's no way for them to tell when we *have*
actually succeeded in saving them.

It'd be good to do better here; details on that in the issue.
For the moment, just stop pretending.