redux types: Drop extras on GlobalThunkAction
These didn't make much sense for GlobalThunkAction in itself,
because they're both redundant when you have a getState that
returns a global state.  We've had them there in order to let
GlobalDispatch be a subtype of Dispatch (and ThunkAction a subtype
of GlobalThunkAction), in order to smooth the process of sorting
out which code should get which of those.

But that subtype relationship won't ultimately make sense either.
It can only work as long as the getState that gets passed to a
GlobalThunkAction is one that a ThunkAction can accept as its own
getState parameter... i.e. only while `() => GlobalState` is a
subtype of the `() => PerAccountState` that a ThunkAction
expects... i.e. only while GlobalState is a subtype of
PerAccountState.  Which is true for now, where there's really just
one per-account state at a time (for the active account), but can't
work in a world with a real multi-account schema with per-account
states for several accounts.

So, we may as well drop these extraneous extras now.  That completes
the two-way distinction between Dispatch and GlobalDispatch, by
making GlobalDispatch no longer a subtype of Dispatch.

(Later in this series of PRs, we'll also make GlobalState no longer
a subtype of PerAccountState, as part of our preparation for
actually changing the schema.)