persist [nfc]: List just keys at first, subtrees a bit later
This is NFC because even though we're deferring some `state[key]`
lookups across an await, we're making them on the same `state` object
(which we have a direct reference to, rather than looking up with a
new getState call each time), and it's assumed not to be mutated.

In itself this version isn't really any more or less simple than
before; but we'll make another change shortly that emphasizes a list
of keys, and this simplifies that change.