getMessageListElements [nfc]: Remove unused header style marker 'none'
When we're inside the `showRecipientHeader` conditional, we already
know that the narrow is not a topic narrow or a PM narrow. So we can
have those be error cases.

I think this is preferable to `caseNarrowPartial` because we don't
want this to start giving errors if other types of interleaved
narrows are added to the `Narrow` type. We just want to specifically
exclude topic and PM narrows in the list of cases we know are
possible.