html/header [nfc]: Remove a dubious line returning the empty string
(Using -b will help with viewing this diff.)

The `header` function is supposed to take a message-list element
descriptor and give the appropriate HTML string to set up the DOM
element in the WebView.

It's important that we establish a 1:1 correspondence between the
"message-list elements" on the RN side and their corresponding DOM
elements in the WebView.

Therefore, that `return ''` line was looking pretty suspicious -- we
don't want a message-list element on the RN side to not have a
corresponding DOM node in the WebView. So, get rid of that line,
after putting in the necessary invariants.