account [nfc]: Rename getRealmUrl, from getCurrentRealm
Saying "current" isn't very helpful here -- it seems to be pointing
at this being about the active account, but (a) that's not a term we
normally use for that meaning, and (b) we normally don't mention that
at all for selectors, almost all of which are for the active account.

So cut that.  Add "URL" instead: it's more specific about what this
produces (not everything about the realm, but just its URL), and also
keeps it distinct from the `getRealm` in `directSelectors`, which
just returns the `realm` subtree of our state.