webview [nfc]: Have getMessageIdFromElement take an Element, and rename
Surprisingly, in most cases Flow has already been convinced that
we're passing Elements; we fixed one easy case in a recent commit.

This will be helpful because we want to reimplement the tiny helper
getMessageNode so that it uses `.closest('msglist-element')`, to
take advantage of our new 'msglist-element' HTML class. And
`.closest` is a method that exists on Element but not Node.