app bar: Give buttons the proper size and spacing
This adjusts the size of the app bar's buttons, and the
horizontal layout of all the buttons and the title, to
match the Material Design spec:
  https://material.io/components/app-bars-top#specs

The motivating issue was that when there are two "action item"
buttons at the end of the app bar, the icons were too widely
spaced apart.  We fix that by giving NavButton the right size.

Together with appropriate padding in ChatNavBar, that's enough
to give all the action-item buttons the exact desired layout.

Other than those action items, the other caller of NavButton is
NavBarBackButton.  With appropriate padding in each of its callers,
that gets laid out correctly too.

Finally, the titles.  Material calls for 32px between the
back-button icon, when there is one, and the start of the title.
At first glance that feels like too much... but after looking
around at some actual screens, I think it isn't -- it still takes
an awfully long stream name to run out of room.  So go ahead and
follow that.