outbox: Make stream_id required on StreamOutbox
We started supplying this property in 0c251bfe0 / #5000.  That went
out to users a few weeks ago, in v17.171 which rolled out 2021-09-23,
so by this point the bulk of our users have upgraded to that version.
Now we mark the property as required, and drop any old outbox messages
that lack it in the unlikely case that there are any.

This completes #3998 because we've already done the same thing for
`sender_id`, in a1fad7ca9 / #4304 and then 60847c9ef / #4667.

Fixes: #3998