autocomplete [nfc]: Accept a null narrow in TopicAutocomplete
This lets us have this widget around when the stream comes from an
input widget which might pass through states where it doesn't have
an actual stream selected -- in particular, ShareToStream, where
the stream name comes from a text input.

This code was already implicitly accepting narrows that weren't an
actual stream, because at present a stream `Narrow` has just a
stream name, and ShareToStream would cheerfully pass a narrow with
the user's arbitrary text as the stream name.  Both of the places
where this code uses the narrow -- getTopicsForNarrow and
fetchTopicsForStream -- silently ignore such narrows.  So this
really just gives us a way to make that explicit, by passing `null`
for the narrow instead.

We'll need that explicitness when we start having stream IDs in our
narrows, and expecting them to be the IDs of real actual streams.