stream [nfc]: Use message.stream_id in a few more places
I went searching for calls to streamNameOfStreamMessage, as likely
places where we could switch to a stream ID now that #3998 is done.

These are the sites that were easy to switch over.  There remain a
few that are showing the name in the UI (so legitimately should be
using the name), or constructing a Narrow (which is a known cluster
of cases that will require its own focused effort), as well as some
that will take a bit more work for miscellaneous other reasons.