generics tests [nfc]: Give alternate angle on variance short circuit
This had said that the check Flow is missing here is at generic-type
definition time.  But on further reflection (and Chris raised this in
review) I think one could see either that check, or a check when
applying variance, without short-circuiting, or both, as what Flow
ought to be doing here.  Which way seems natural depends on how one
makes precise what a polymorphic type with bounds on its parameters,
or a variance marking on a type parameter, is supposed to mean.

So instead of "The bug here is", describe both versions.

Either way, though, the check at definition time is the one that's
practical for us to try to apply manually, so that's the one we'll go
for as a workaround.

For that reason I also leave in place the references to this issue
at the two definitions where we fixed the variance; we can keep
those simple (well, relatively so), with just the side of this
that's actionable in those contexts, and leave the complexity for
the longer prose here which they point to.