msglist [nfc]: Drop redundant, not-quite-right type annotation
We have a libdef describing the WebView component this function
gets passed as a callback prop to, and it has nice clear types.
The type passed to this callback as its argument is not in fact
WebViewNavigation; it's a different type, which resembles that
but has an additional property.

Rather than go import and use the name the libdef invents for
this callback-argument type, we can just let Flow infer the type.
That's its job, which it is happy to do within any module (we're
not exporting this function.)  Semantically this function's type
is really "whatever makes an appropriate onShouldStartLoadWithRequest
callback for WebView", and that's exactly what we get if we just
let Flow do its thing.

(Even if we were exporting this function from a module so that we
did have to write down a type, I'd want to find a way to write it
as that semantic description: should be possible by starting from
`typeof WebView` and using `React.ElementConfig`.)