android sharing [nfc]: Rename and javadoc the shim activity
This is an activity, not an intent, so its name should end with
"Activity" and not with "Intent".

We can also name it a bit more for what its purpose is at a
product level; the fact that it gets intents whose action is
SEND (and SEND_MULTIPLE) is expressed more clearly in the
relevant code in the manifest anyway.  Also add a bit of javadoc
to explain what's going on and why it doesn't look like it's
doing much.