caught-up: Drop legacy inference for servers pre-1.8
Huzzah!  Look at that negative diffstat.

Also make a note next to where we record `minSupportedVersion`, in
ServerCompatBanner.js.  We may at some point want to have the app
outright refuse to operate on sufficiently ancient servers, where
it's just not going to be a good experience.  Until then, it seems
helpful to track what we know about what the consequences are
likely to be (though actually *testing* on such ancient servers
would be significantly more effort and not worth it.)