notif [nfc]: Give getAccountFromNotificationData accounts, not identities
When we added this function in 056930c60, we may have specifically
chosen Identity objects over Account objects, to avoid adding a
place where we pass around auth credentials. Hmm.

Still, we've recently started storing the user ID in Account objects
(4fdefb09b), and we have a TODO in this function for using it.

The Identity type doesn't have the user ID, and it's not currently
convenient to add it there. That's because the Identity type is
based on the Auth type (and we make Identity objects from Auth
objects), and the Auth type doesn't have the user ID. It's not super
convenient to add the user ID to the Auth type because there are
places where we make Auth objects but we don't have a user ID, like
`authFromCallbackUrl` in src/start/webAuth.js. As Greg points
out [1], those places are obstacles to #3764, "Convert from emails
to user IDs to identify users".